Reduce modify_contract gas usage
