+++ Cargo.toml
@@ -30,7 +30,6 @@ overflow-checks = true
 [dependencies]
 cosmwasm-std = { version = "1.3.3" }
 cw-storage-plus = { version = "1.1.0" }
-cosmwasm-storage = { version = "1.3.3", features = ["iterator"] }
 provwasm-std = { version = "1.1.2" }
 rust_decimal = "1.29.0"
 schemars = "0.8.11"
