+++ src/tests/execute/expire_bid_tests.rs
@@ -193,6 +193,7 @@ mod expire_bid_tests {
               \"denom\": \"quote_1\",
               \"total_supply\": \"1000\",
               \"marker_type\": \"restricted\",
+              \"allow_forced_transfer\": false,
               \"supply_fixed\": false
             }";
